[fix] view updater(tile by tile)

+++ src/main/java/org/gamedevs/clashroyale/model/game/droppable/objects/soldiers/Soldier.java
@@ -162,7 +162,7 @@ public abstract class Soldier extends GameObject {
             path = pathFinder.getPath();
         }
         if (state == GameObjectState.MOVING) {
-            if((currentFrame - initialFrame) % ((int)(speed * 10)) != 0)
+            if((currentFrame - initialFrame) % ((speed * 10)) != 0)
                 return;
             Console.getConsole().printTracingMessage(ConsoleColor.GREEN_BOLD + "In moving state");
             if(closestTargetTile == null){
